# [opt] Optimizing Shoup's MulMod in Apple M1/M2 chip.