@tlepoint 2024-03-26
@sarah-ek Is this an optimization you would consider for `concrete-ntt`?