@IceTDrinker 2024-11-08
we will be renaming concrete-ntt to tfhe-ntt and moving it to the tfhe-rs repo, is this still relevant or have newer versions of the compiler removed that "weird" behavior @fionser @tlepoint ?